+++ test.sh
@@ -18,3 +18,6 @@ bash nvim/test_nvim.sh
 
 echo "Testing bash shell integration"
 bash shell/bashrc_test.sh
+
+echo "Testing build.sh"
+bash build_test.sh